+++ set-function/xor-convolution.hpp
@@ -4,7 +4,7 @@ template <class T> vector<T> xor_convolution(vector<T> a, vector<T> b) {
     assert(a.size() == b.size());
     fast_hadamard_transform(a);
     fast_hadamard_transform(b);
-    for(int i = 0; i < a.size(); i++)
+    for(int i = 0; i < (int)a.size(); i++)
         a[i] *= b[i];
     fast_hadamard_transform(a, true);
     return a;
