+++ test/library_checker/number_theory/enumerate_primes.test.cpp
@@ -5,7 +5,7 @@ void solve() {
     LL(n, a, b);
     auto primes = enumerate_primes(n);
     vector<int> ans;
-    for(int i = b; i < primes.size(); i += a) {
+    for(int i = b; i < (int)primes.size(); i += a) {
         ans.emplace_back(primes[i]);
     }
     print(primes.size(), ans.size());
